fix(server): Fix error when populating assets
When running the populate script, I was getting an ERR_STREAM_WRITE_AFTER_END error each time. It seems related to https://github.com/jaydenseric/graphql-upload/issues/131 but unfortunately this is a transitive dependency of apollo-server-express and I can't directly upgrade to the latest version of graphql-upload. This work-around seems to fix things though.