feat(core): Allow custom ApolloServerPlugins to be specified
For certain types of data transformation, express plugins and Nestjs interceptors are not sufficient since they occur at the wrong point in the request/response pipeline. This is the case with the built-in ID-transformation code. For these problems, ApolloServerPlugins provide very fine-grained control over the request/response, with full access to the GraphQL schema and document. This can in theory enable very advanced use-cases such as implementing custom GraphQL directives.

Closes #210