feat(create): Simplify create steps - remove JS option
Offering JS as an option doesn't make much sense - sure the initial project will build, but actually
creating anything non-trivial in JS without decorators support is, in practical terms, painful.

We don't provide _any_ documentation on e.g. how to define a plugin with pure JS - it's just not
practical. So offering the JS option on project creation is a red herring and is also dishonest
in the sense that it seems to suggest we support it.