feat(core): Implement Regions & support for Provinces
Relates to #76

BREAKING CHANGE: A new `Region` entity has been introduced, which is a base class for `Country` and
the new `Province` entity. The `Zone.members` property is now an array of `Region` rather than
`Country`, since Zones may now be composed of both countries and provinces. If you have defined
any custom fields on `Country`, you'll need to change it to `Region` in your custom fields config.